+++ packages/web3-eth/src/web3_eth.ts
@@ -1590,11 +1590,7 @@ export class Web3Eth extends Web3Context<Web3EthExecutionAPI, RegisteredSubscrip
 		args?: ConstructorParameters<RegisteredSubscription[T]>[0],
 		returnFormat: ReturnType = DEFAULT_RETURN_FORMAT as ReturnType,
 	): Promise<InstanceType<RegisteredSubscription[T]>> {
-		const subscription = (await this.subscriptionManager?.subscribe(
-			name,
-			args,
-			returnFormat,
-		)) as InstanceType<RegisteredSubscription[T]>;
+		const subscription = await this.subscriptionManager?.subscribe(name, args, returnFormat);
 		if (
 			subscription instanceof LogsSubscription &&
 			name === 'logs' &&